+++ src/pages/how-it-works.js
@@ -1,9 +1,9 @@
 import PropTypes from "prop-types";
 import React from "react";
 
+import MetaTextBlock from "@/pesayetu/components/MetaTextBlock";
 import Hero from "@/pesayetu/components/OtherHero";
 import Page from "@/pesayetu/components/Page";
-import TwoColumnLayout from "@/pesayetu/components/TwoColumnLayout";
 import formatBlocksForSections from "@/pesayetu/functions/formatBlocksForSections";
 import getPostTypeStaticProps from "@/pesayetu/functions/postTypes/getPostTypeStaticProps";
 
@@ -11,7 +11,7 @@ export default function Home({ blocks, ...props }) {
   return (
     <Page {...props}>
       <Hero {...blocks?.otherHero} />
-      <TwoColumnLayout {...blocks?.twoColumnLayout} />
+      <MetaTextBlock {...blocks?.twoColumnLayout} />
     </Page>
   );
 }
